// projekat_pia/frontend/src/app/model/canvas.ts
export class Canvas{
    brSoba: number;
    brVrata: number;
    koordinateSoba: {
        x: number;
        y: number;
        width: number;
        height: number;
        color: string;
    }[];
    koordinateVrata: {
        x: number;
        y: number;
        width: number;
        height: number;
    }[];
}